refactor: rename GroupSingle
